requirements.txt
pillow
flask
tensorflow
werkzeug
cassandra-driver










